resume_parser/requirements.txt
# Core dependencies for resume_parser
PyPDF2>=3.0.1
python-docx>=1.1.0
google-genai>=0.1.0
python-dotenv>=1.0.0